Split components into different files
